requirements.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
airportsdata==20250909
annotated-types==0.7.0
anyio==4.11.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
cloudpickle==3.1.1
comm==0.2.3
contourpy==1.3.3
cramjam==2.11.0
cycler==0.12.1
datasets==4.2.0
debugpy==1.8.17
decorator==5.2.1
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.17.0
dotenv==0.9.9
et_xmlfile==2.0.0
executing==2.2.1
fastparquet==2024.11.0
filelock==3.19.1
fonttools==4.60.0
frozenlist==1.8.0
fsspec==2025.9.0
genson==1.3.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.10
inquirerpy==0.3.4
instructor==1.11.3
interegular==0.3.3
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
lark==1.3.0
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.3
openai==1.109.1
openpyxl==3.1.5
outlines_core==0.1.26
packaging==25.0
pandas==2.3.2
parso==0.8.5
patsy==1.0.1
pexpect==4.9.0
pfzy==0.3.4
pillow==11.3.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.1.0
psycopg2==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pycountry==24.6.1
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2025.9.18
requests==2.32.5
rich==14.2.0
rpds-py==0.27.1
safetensors==0.6.2
scipy==1.16.2
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
stack-data==0.6.3
statsmodels==0.14.5
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.22.1
torch==2.9.0
torchvision==0.24.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
widgetsnbextension==4.0.14
xxhash==3.6.0
yarl==1.22.0
